# chore(optimization): Tiny optimization for pointsInternal