--- a/src/core/geometry/qgsellipse.cpp
+++ b/src/core/geometry/qgsellipse.cpp
@@ -244,12 +244,12 @@ void QgsEllipse::pointsInternal( unsigned int segments, QVector<double> &x, QVec
   const double sinAzimuth = std::sin( azimuth );
   for ( double it : t )
   {
-    *xOut++ = centerX +
-              mSemiMajorAxis * std::cos( it ) * cosAzimuth -
-              mSemiMinorAxis * std::sin( it ) * sinAzimuth;
-    *yOut++ = centerY +
-              mSemiMajorAxis * std::cos( it ) * sinAzimuth +
-              mSemiMinorAxis * std::sin( it ) * cosAzimuth;
+    const double cosT{ std::cos( it ) };
+    const double sinT{ std::sin( it ) };
+    *xOut++ = centerX + mSemiMajorAxis * cosT * cosAzimuth -
+              mSemiMinorAxis * sinT * sinAzimuth;
+    *yOut++ = centerY + mSemiMajorAxis * cosT * sinAzimuth +
+              mSemiMinorAxis * sinT * cosAzimuth;
     if ( zOut )
       *zOut++ = centerZ;
     if ( mOut )