@uclaros 2025-01-14
Are those results on a _release_ build?
I'd expect the compiler/cpu to render this kind of optimization unnecessary.