@lbartoletti 2025-01-21
> Are those results on a release build?

@uclaros I'm live only in a Release mode :wink: 

> > I'd expect the compiler/cpu to render this kind of optimization unnecessary.
> 
> I'm not sure the compiler (before C++26 which will declare them constexpr: https://en.cppreference.com/w/cpp/numeric/
math/cos) can infer that the output of std::cos() doesn't change for a given output. But instead of creating an array, which takes time, I'd suggest to just to just create temporary c = std::cos(t[i]) and s= std::sin(t[i]) variables in the loop.
> 
> > reducing the number of expensive trigonometric function calls in the main loop.
> 
> hum, I would expect that your optimization would just divide the time by 2, not by 10. Did you look at the dissembled code to understand why you get such perf gain ? Which compiler do you use ? Maybe OpenMP is used and the array computation is multithreaded ? In any case adding a comment in the code explaining why the optimization works and under which circumstances would make later maintenance easier.

@rouault I did a more extensive test, which you can find [here](https://github.com/lbartoletti/lbartoletti.github.io/tree/master/archives/qgis/ellipse_benchmark). It's not real QGIS anymore, but I think it's still representative.

@ptitjano  and I have been talking about testing OpenMP on QGIS for several years now, and I've been looking a lot at what's being done on GRASS. We're thinking of proposing a QEP within the year, but first we need to do some testing (and learn OpenMP properly). 

In the meantime, for this test, you can see that depending on the number of threads used, OpenMP is interesting around 256 vertices. In any case, it's better to use method 2 with the creation of cos/sin in the loop, as you suggest, and not to pre-calculate it as I did.

So we have choice between:

OpenMP ready
```cpp
for (int i = 0; i < static_cast<int>(segments); ++i) {
            const double cosT{ std::cos(t[i]) };
            const double sinT{ std::sin(t[i]) };
            x[i] = centerX + mSemiMajorAxis * cosT * cosAzimuth -
                mSemiMinorAxis * sinT * sinAzimuth;
            y[i] = centerY + mSemiMajorAxis * cosT * sinAzimuth +
                mSemiMinorAxis * sinT * cosAzimuth;
            if (hasZ) z[i] = centerZ;
            if (hasM) m[i] = centerM;
        }
```


Actual with cos/sin calculated  only one time.
```cpp
 for (double it : t) {
            const double cosT{ std::cos(it) };
            const double sinT{ std::sin(it) };
            *xOut++ = centerX + mSemiMajorAxis * cosT * cosAzimuth -
                mSemiMinorAxis * sinT * sinAzimuth;
            *yOut++ = centerY + mSemiMajorAxis * cosT * sinAzimuth +
                mSemiMinorAxis * sinT * cosAzimuth;
            if (zOut) *zOut++ = centerZ;
            if (mOut) *mOut++ = centerM;
        }
```