@lbartoletti 2025-01-29
> > @rouault I did a more extensive test, which you can find [here](https://github.com/lbartoletti/lbartoletti.github.io/tree/master/archives/qgis/ellipse_benchmark). It's not real QGIS anymore, but I think it's still representative.
> 
> @lbartoletti I'm confused if you put this standalone benchmark just to answer my question... So please bear with my "too long didn't read" attitude here. 

No worries, it was just an interesting test for us!

> Basically I was just curious about the reason why this PR gives such perf boost. In short, is OpenMP implicitly used by your optimizer... ?

No, I don't think.