@lbartoletti 2025-02-17
> @lbartoletti
> 
> > But instead of creating an array, which takes time, I'd suggest to just to just create temporary c = std::cos(t[i]) and s= std::sin(t[i]) variables in the loop.
> 
> Can you do this so that this change can be merged?

done. Thanks